removing unwanted files
