+++ src/Kmd.Logic.Audit.Client.SerilogLargeAuditEvents/SerilogLargeAuditEventClient.cs
@@ -5,28 +5,22 @@ namespace Kmd.Logic.Audit.Client.SerilogLargeAuditEvents
 {
     public class SerilogLargeAuditEventClient : IAudit, IDisposable
     {
-        private readonly Logger eventhubLogger;
-        private readonly Logger blobLogger;
+        private readonly Logger customSinkLogger;
         private readonly bool disposeLogger;
-        private readonly SerilogLoggerAudit audit;
-        private readonly SerilogBlobLoggerAudit auditToBlob;
+        private readonly SerilogCustomSinkLoggerAudit auditToCustomSink;
 
         public SerilogLargeAuditEventClient(SerilogLargeAuditEventClientConfiguration config)
         {
-            this.eventhubLogger = config.CreateEventhubConfiguration().CreateLogger();
-            this.blobLogger = config.CreateBlobConfiguration().CreateLogger();
+            this.customSinkLogger = config.CreateBlobConfiguration().CreateLogger();
             this.disposeLogger = true;
-            this.audit = new SerilogLoggerAudit(this.eventhubLogger);
-            this.auditToBlob = new SerilogBlobLoggerAudit(this.blobLogger);
+            this.auditToCustomSink = new SerilogCustomSinkLoggerAudit(this.customSinkLogger);
         }
 
         private SerilogLargeAuditEventClient(Logger logger, bool disposeLogger)
         {
-            this.eventhubLogger = logger;
-            this.blobLogger = logger;
+            this.customSinkLogger = logger;
             this.disposeLogger = disposeLogger;
-            this.audit = new SerilogLoggerAudit(this.eventhubLogger);
-            this.auditToBlob = new SerilogBlobLoggerAudit(this.blobLogger);
+            this.auditToCustomSink = new SerilogCustomSinkLoggerAudit(this.customSinkLogger);
         }
 
         public static SerilogLargeAuditEventClient CreateCustomized(Logger logger, bool disposeLogger = true)
@@ -36,21 +30,19 @@ namespace Kmd.Logic.Audit.Client.SerilogLargeAuditEvents
 
         public void Write(string messageTemplate, params object[] propertyValues)
         {
-            //this.audit.Write(messageTemplate, propertyValues);
-            this.auditToBlob.Write(messageTemplate, propertyValues);
+            this.auditToCustomSink.Write(messageTemplate, propertyValues);
         }
 
         public IAudit ForContext(string propertyName, object value, bool captureObjectStructure = false)
         {
-            return this.audit.ForContext(propertyName, value, captureObjectStructure);
+            return this.auditToCustomSink.ForContext(propertyName, value, captureObjectStructure);
         }
 
         protected virtual void Dispose(bool disposing)
         {
             if (disposing && this.disposeLogger)
             {
-                this.eventhubLogger.Dispose();
-                this.blobLogger.Dispose();
+                this.customSinkLogger.Dispose();
             }
         }
